Refactoring: encapsulating data and behavior.
